move classes from root into sub packages
