+++ src/main/java/org/betonquest/betonquest/BetonQuest.java
@@ -42,6 +42,9 @@ import org.betonquest.betonquest.database.MySQL;
 import org.betonquest.betonquest.database.SQLite;
 import org.betonquest.betonquest.database.Saver;
 import org.betonquest.betonquest.item.QuestItemHandler;
+import org.betonquest.betonquest.listener.CustomDropListener;
+import org.betonquest.betonquest.listener.JoinQuitListener;
+import org.betonquest.betonquest.listener.MobKillListener;
 import org.betonquest.betonquest.logger.DefaultBetonQuestLoggerFactory;
 import org.betonquest.betonquest.logger.HandlerFactory;
 import org.betonquest.betonquest.logger.PlayerLogWatcher;
