+++ src/main/java/org/betonquest/betonquest/command/CancelQuestCommand.java
@@ -1,8 +1,8 @@
 package org.betonquest.betonquest.command;
 
-import org.betonquest.betonquest.Backpack;
-import org.betonquest.betonquest.Backpack.DisplayType;
 import org.betonquest.betonquest.config.PluginMessage;
+import org.betonquest.betonquest.feature.Backpack;
+import org.betonquest.betonquest.feature.Backpack.DisplayType;
 import org.betonquest.betonquest.util.PlayerConverter;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
