+++ src/main/java/org/betonquest/betonquest/item/QuestItemHandler.java
@@ -1,9 +1,9 @@
 package org.betonquest.betonquest.item;
 
-import org.betonquest.betonquest.Journal;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.config.Config;
 import org.betonquest.betonquest.data.PlayerDataStorage;
+import org.betonquest.betonquest.feature.journal.Journal;
 import org.betonquest.betonquest.util.PlayerConverter;
 import org.betonquest.betonquest.util.Utils;
 import org.bukkit.GameMode;
