+++ src/main/java/org/betonquest/betonquest/quest/event/journal/JournalChanger.java
@@ -1,6 +1,6 @@
 package org.betonquest.betonquest.quest.event.journal;
 
-import org.betonquest.betonquest.Journal;
+import org.betonquest.betonquest.feature.journal.Journal;
 
 /**
  * Defines changes to be done to a journal.
