+++ src/main/java/org/betonquest/betonquest/quest/event/journal/package-info.java
@@ -1,5 +1,5 @@
 /**
  * {@link org.betonquest.betonquest.api.quest.event.Event Event} implementation for actions
- * related to the BetonQuest {@link org.betonquest.betonquest.Journal Journal}.
+ * related to the BetonQuest {@link org.betonquest.betonquest.feature.journal.Journal Journal}.
  */
 package org.betonquest.betonquest.quest.event.journal;
