+++ src/test/java/org/betonquest/betonquest/quest/event/journal/JournalEventTest.java
@@ -1,9 +1,9 @@
 package org.betonquest.betonquest.quest.event.journal;
 
-import org.betonquest.betonquest.Journal;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.data.PlayerDataStorage;
 import org.betonquest.betonquest.database.PlayerData;
+import org.betonquest.betonquest.feature.journal.Journal;
 import org.betonquest.betonquest.quest.event.NotificationSender;
 import org.betonquest.betonquest.util.PlayerConverter;
 import org.bukkit.entity.Player;
